requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
python-dateutil==2.8.2
openpyxl==3.1.2
akshare>=1.14.0
requests>=2.31.0
